Rename files to avoid pytest conflicts
