+++ ics/event.py
@@ -12,8 +12,8 @@ from .organizer import Organizer
 from .parse import Container
 from .types import ArrowLike
 from .utils import (get_arrow, uid_gen)
-from ics.parsers.event import EventParser
-from ics.serializers.event import EventSerializer
+from ics.parsers.event_parser import EventParser
+from ics.serializers.event_serializer import EventSerializer
 
 
 class Geo(NamedTuple):
